refactor(hooks): move usePwa from lib/ to hooks/
usePwa is a plain React hook (useSyncExternalStore) with no lib-level
coupling, yet it was the only use* living under lib/. src/hooks/ is the
established home for the project's custom hooks, so move it there for
consistency. Pure relocation + import-path update; no behaviour change.
