+++ src/components/Nav.tsx
@@ -13,7 +13,7 @@ import { useEffectiveMount } from "@/hooks/useMountParam";
 import { useBreakpoint } from "@/hooks/useBreakpoint";
 import { mountToUrlSegment } from "@/lib/mount";
 import { useNav } from "@/context/NavContext";
-import { usePwa } from "@/lib/usePwa";
+import { usePwa } from "@/hooks/usePwa";
 import { track } from "@/lib/analytics";
 import { cn } from "@/lib/utils";
 import { MENU_POPUP_CLS } from "@/lib/ui-tokens";
